Move `FeedViewAdapter` to its own file
