+++ EssentialFeediOS/FeedUI/Composers/FeedViewAdapter.swift
@@ -1,0 +1,29 @@
+//
+//  FeedViewAdapter.swift
+//  EssentialFeediOS
+//
+//  Created by Jair Moreno Gaspar on 05/01/21.
+//  Copyright © 2021 Jair Moreno G. All rights reserved.
+//
+
+import UIKit
+
+final class FeedViewAdapter: FeedView {
+
+    private weak var controller: FeedViewController?
+    private let loader: FeedImageDataLoader
+    
+    init(controller: FeedViewController, loader: FeedImageDataLoader) {
+        self.controller = controller
+        self.loader = loader
+    }
+    
+    func display(_ viewModel: FeedViewModel) {
+        controller?.tableModel = viewModel.feed.map { model in
+            let adapter = FeedImageDataLoaderPresentationAdapter<WeakRefVirtualProxy<FeedImageCellController>, UIImage>(model: model, imageLoader: loader)
+            let view = FeedImageCellController(delegate: adapter)
+            adapter.presenter = FeedImagePresenter(view: WeakRefVirtualProxy(view), imageTransformer: UIImage.init)
+            return view
+        }
+    }
+}
